Fix PC resolution in the pipelined processor
In a pipelined processor the "true" value of the PC register could be
anything with regards to the PC of an instruction in the execute
stage. As a result the _fetch_ PC of the instruction plus four is used
as per convention. This patch fixes the standard opcode to implement
this behavior.