Use test/is rather than booleans
Clean up the test case predicates to use clojure.test/is rather than
assert or combinational logic. This allows for the specific failure to
be reported to the programmer rather than a simple indication of the
failure's existence.

Also mixed in here is a whitespace "cleanup" and a rewrite of the
run-test fn to use the new t/ alias for clojure.test/