Patch encoding tests
This patch fixes an issue where the byte encoded tests had :opcodes
fields with sequences of byte encoded instructions rather than the
required mapping from addresses to instructions.

This patch also fixes a related issue with the the encoding tests,
wherein the encoded mapping would be compared to the ordered
instruction sequence.

All tests now pass, Batbridge is clear for 0.1.0