Fix a series of bugs related to register vals
Previously, there were two publicly usable functions, get-register and
register->val. Problem with this is that get-register actually read a
value out of the registers map, and register->val gets some of its
possible results from arguments. Critically, it's meaningless to read
the immediate register from the registers array since it's a source
defined on a per-op basis.

This patch hides get-register and refactors all of its callsites to use
a form of register->val which will complain loudly if a user somehow
attempts to reference the immediate register without providing it.