Fix bug using the imm as a source for load
Most of the ops on Batbridge are of the format two arguments, one
destination. Except for `st`, which takes three arguments. This exposed
an implementation bug where a user attempting to reference the immediate
field from a `st` instruction, say as the data source, would get a
"real" register file read of 0 as a result rather than the value of the
immediate field.