WIP pipelines less broken but needs major work
The fundimental problem is that it's not at all clear where the state of
a given "stage" is, and where the result of a given "stage" is. At
present, there's one :fetch, :decode and soforth all of which represent
the result of the last time that stage was evaluated. This works, but is
inelegant. So. 1. Fix that. :result/fetch, :result/decode etc or
something. Make the intent of those values clear.

The other thing is that all of Batbridge and Toothpick predates my work
with/on smart constructors. So I'm running into all these neat cases
(previously dealt with by core.typed before I threw that out) where if
the previous stage generated no value or a pipeline bubble then we have
a problem. This needs to get cleaned up either by creating a bunch of
deftags to manage pipeline results, and by adding pre/post conditions on
every stage so that there's a contract I can program to.

Blarg. Usual Clojure refactoring pain applies.