Branch predict all instructions
I added this code so that only nominal branch instructions were
predicted while debugging the branch predicted pipeline. The problem
with this is any instruction _could_ impact the PC. We don't know if it
will until we decode it and we need to predict now. So we just make a
prediction for everything.

This turns out to be a huge performance win because it means that
unconditional jumps (as in the bottom of a loop) encoded as add or
subtract operations on the PC will now be predicted. With prediction
limiting in place, these instructions were not predicted and the bottom
of a loop cost a pipeline flush every time.